Initialize Event Grid extension (#1)
Added all of the basic boilerplate files needed for an extension as well as basic tree functionality including OpenInPortal/Delete/Refresh/LoadMore.

Event Grid actually has two separate resource types in Azure, so there are two tree views. This might be a little annoying, but it's consistent with the Portal (which has two separate views) and the user can always hide a window if they're not using it. The two types are "topics" and "event subscriptions". Here's a link to the overview of Event Grid if you want more info: https://docs.microsoft.com/azure/event-grid/overview