Update README for Multi Tenant Accounts and Auth (#38)
It took me a second to figure out how to sign-on with my multi-tenant account. I had to figure out that the azureeventgrid extension used this extension. Then I saw the tenant setting and put everything together.

Having the settings listed here combined with documentation on the other end should help speed that whole process up.